refactor: desk navigation refactoring
