+++ src/components/common/navigation/MobileMenu/index.tsx
@@ -7,7 +7,7 @@ import {
   NavIconWrapper,
   ShortNavigation,
   NavigationWapper,
-} from '../styled';
+} from '../DeskNavigation/styled';
 import { useRecoilState } from 'recoil';
 import { MENU_KEY, menuState } from '../../../../store/menu';
 import MobileMenuCategory from '../MobileMenuCategory';
