+++ src/components/common/navigation/WideNavCategory/index.tsx
@@ -1,10 +1,15 @@
 import React, { useEffect, useState } from 'react';
-import { NavTask, NavTaskWrapper, StyledLink, WideNavigation } from '../styled';
+import {
+  NavTask,
+  NavTaskWrapper,
+  StyledLink,
+  WideNavigation,
+} from '../DeskNavigation/styled';
 import { useLocation } from 'react-router';
 import { useRecoilState } from 'recoil';
 import { MENU_KEY, menuState } from '../../../../store/menu';
 import { useNavigate } from 'react-router-dom';
-import { navigationData } from '../index';
+import { navigationData } from '../DeskNavigation';
 
 const WideNavCategory = ({ routeStyle }: any) => {
   const navigate = useNavigate();
